Use python3 for tensorflow docker containers
Simple modification to publish.sh to use python3 for all new docker
containers. Basically "latest" and "latest-py3" will be the same
image with different names.

The alternative was to remove all the python 3 or 2 logic, but that
logic might come in handy in the future if we need to support
another python release.

Added a message on what image we are building to the logs to help
with debug.