Update for h5py version 2.10.0
h5py version 2.10.0 was released on September 6th, and it requires
the command pkg-config in order to build from source. In every
docker container except for cpu-ppc64le.Dockerfile we had already
installed pkg-config.

The nice thing about this new version, it is able to build with
libhdf5-dev installed without any hacks to find libhdf5-dev or the
.h file. So these hacks have been removed.