Move the base url out of dockerfiles and into the publish script (#132)
The dockerfiles all have duplicate code to determine the url from which to pull
the tensorflow wheel from. In general there are 4 places to pull the wheel file
from:
cpu release builds
gpu release builds
cpu nightly builds
gpu nightly builds

This logic works fine, but it is better to have it in publish.sh one time, and
then pass the BASE_URL into the dockerfile so it can download and install the
correct wheel.

Sometimes we want to publish a docker image using a URL that isn't one of the 4
builds above. The tool doesn't currently handle that. This PR adds a -u (for url)
option to publish.sh to pass in a URL from which the dockerfile should pull the
tensorflow wheel file from.

This allows for special case builds like LMS and TensorFlow 1.15.X builds to be
published in docker without having to clone the branch and hack the scripts.