Create strimzi-kafka-oauth.travis.yml
Added support for ppc64le for the existing travis.yml file in the repo. The build is successful for ppc64le architecture and one of the language is failing on amd64 which was seen in previous build too. Let me know whether the issue should be raised in community for amd64 build failure?