Create archiver_ubi_8.4.sh
Verified build and test of archiver package on ppc64le with master (v3.5.0) and v3.2.0 as updated in the spread sheet.

Create archiver_ubi_8.4.sh

Verified build and test of archiver package on ppc64le with master (v3.5.0) and v3.2.0 as updated in the spread sheet.

Create LICENSE

Delete archiver_ubi_8.4.sh