OSE: Added License's to the build_script repo (#7526)
* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0 to the project

* Create LICENSE

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0 to Docker installation docs

* Add Apache License 2.0 to the project

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Create LICENSE

* Add Apache License 2.0 to gha-script

* Add Apache License 2.0

Added the Apache License 2.0 to the repository.

* Create LICENSE

* Create LICENSE

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0 to the project

* Create LICENSE

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Create LICENSE

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0 to the project

* Create LICENSE

* Create LICENSE

* Add Apache License 2.0

Added Apache License 2.0 to the project.

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Create LICENSE

* Create LICENSE

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Create LICENSE

* Create LICENSE

* Add Apache License 2.0

Added the Apache License 2.0 to the project.

* Create LICENSE

* Create LICENSE

* Create LICENSE

* Create LICENSE